requirements.txt
ultralytics>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
Pillow>=8.0.0
Flask>=2.0.0
waitress>=2.0.0
matplotlib>=3.3.0
numpy>=1.21.0
